# docs/videoapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an auto-generated API Docs.",
        "title": "Video service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "podkidysh2002@gmail.com."
        },
        "version": "1.0"
    },
    "paths": {
        "/tickets": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authorized users can get all video tickets.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Get tickets",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VideoTicket"
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong creadentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authorized users can create tickets for processing uploaded videos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Create new video ticket",
                "parameters": [
                    {
                        "description": "info for video processing",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VideoTicket"
                        }
                    },
                    "401": {
                        "description": "Wrong creadentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Provided media type is not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authorized users can get uploaded videos by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Get video by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Video"
                        }
                    },
                    "401": {
                        "description": "Wrong creadentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/videos": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authorized users can get all uploaded videos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Get videos",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Video"
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong creadentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authorized users can upload their videos",
                "consumes": [
                    "video/webm"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Create new video",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Video"
                        }
                    },
                    "401": {
                        "description": "Wrong creadentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Provided media type is not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/videos/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Authorized users can get uploaded videos by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video"
                ],
                "summary": "Get video by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Video"
                        }
                    },
                    "401": {
                        "description": "Wrong creadentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.CreateTicketRequest": {
            "type": "object",
            "required": [
                "crf",
                "target_format",
                "video_id"
            ],
            "properties": {
                "crf": {
                    "type": "integer"
                },
                "target_format": {
                    "type": "string"
                },
                "video_id": {
                    "type": "string"
                }
            }
        },
        "models.Video": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "video_id": {
                    "type": "string"
                }
            }
        },
        "models.VideoTicket": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "crf": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "target_format": {
                    "type": "string"
                },
                "ticket_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "video_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}